extract-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.18
python-dotenv==1.0.0
markitdown[pdf,docx,xlsx,pptx]>=0.0.1,<1.0.0
